Working on the controllers

+++ final-capstone/back-end/java/src/main/java/com/techelevator/application/jdbcdao/jdbcReviewDao.java
@@ -46,7 +46,7 @@ public class jdbcReviewDao implements reviewDao{
 		review.setDescription(row.getString("description"));
 		review.setRating(row.getInt("rating"));
 		review.setCreateTime(row.getTimestamp("create_date").toLocalDateTime());
-		review.setBeerId(row.getInt("beer_id"));
+		review.setBeerId(row.getLong("beer_id"));
 		review.setUserId(row.getLong("user_id"));
 		return review;
 	}
